requirements.txt
astroid==2.15.6
attrs==23.1.0
awswrangler==3.3.0
boto3==1.28.43
botocore==1.31.43
certifi==2023.7.22
charset-normalizer==3.2.0
dill==0.3.7
dnspython==2.4.2
et-xmlfile==1.1.0
exceptiongroup==1.1.3
idna==3.4
iniconfig==2.0.0
isort==5.12.0
jmespath==1.0.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
lazy-object-proxy==1.9.0
mccabe==0.7.0
numpy==1.25.2
openpyxl==3.1.2
packaging==23.1
pandas==2.1.0
platformdirs==3.10.0
pluggy==1.3.0
pyarrow==13.0.0
pylint==2.17.5
pymongo==4.5.0
pytest==7.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.2
s3transfer==0.6.2
six==1.16.0
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
wrapt==1.15.0
